requirements.txt
markdownify
SpeechRecognition
pyaudio
pydub
openai-whisper
pyChatGPT @ git+https://github.com/hesiod-au/pyChatGPT@v0.3-vanilla
pyperclip
